requirements.txt
certifi==2020.11.8
chardet==3.0.4
greenlet==1.0a1
idna==2.10
playwright==0.162.2
psycopg2-binary==2.8.6
pyee==8.1.0
requests==2.25.0
selenium==3.141.0
selenium-stealth==1.0.6
TikTokApi==3.8.3
typing-extensions==3.7.4.3
urllib3==1.26.2
